+++ src/Traits/Configuration/CustomisationsConfiguration.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Rappasoft\LaravelLivewireTables\Traits\Configuration;
+
+trait CustomisationsConfiguration
+{
+}