+++ src/Traits/Helpers/CustomisationsHelpers.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Rappasoft\LaravelLivewireTables\Traits\Helpers;
+
+trait CustomisationsHelpers
+{
+}