+++ src/Views/Columns/LinkColumn.php
@@ -21,7 +21,7 @@ class LinkColumn extends Column
 
     protected mixed $attributesCallback = null;
 
-    public function __construct(string $title, string $from = null)
+    public function __construct(string $title, ?string $from = null)
     {
         parent::__construct($title, $from);