+++ src/Views/Filter.php
@@ -9,7 +9,7 @@ abstract class Filter
 {
     use IsFilter;
 
-    public function __construct(string $name, string $key = null)
+    public function __construct(string $name, ?string $key = null)
     {
         $this->name = $name;
 
@@ -24,7 +24,7 @@ abstract class Filter
     /**
      * @return static
      */
-    public static function make(string $name, string $key = null): Filter
+    public static function make(string $name, ?string $key = null): Filter
     {
         return new static($name, $key);
     }