+++ tests/TestCase.php
@@ -72,14 +72,23 @@ class TestCase extends Orchestra
                 ['id' => 4, 'pet_id' => 2, 'veterinary_id' => 3],
             ]);
         }
+        $this->setupBasicTable();
+    }
 
+    protected function setupBasicTable()
+    {
+        $view = view('livewire-tables::datatable');
         $this->basicTable = new PetsTable();
+        $this->basicTable->bootComponentUtilities();
         $this->basicTable->boot();
+        $this->basicTable->bootWithData();
+        $this->basicTable->bootWithColumns();
+        $this->basicTable->bootWithColumnSelect();
+        $this->basicTable->bootWithSecondaryHeader();
         $this->basicTable->booted();
-        $this->basicTable->callHook('rendering');
-        $this->basicTable->callTraitHook('rendering');
+        $this->basicTable->renderingWithData($view, []);
+        $this->basicTable->renderingWithPagination($view, []);
         $this->basicTable->render();
-        $this->basicTable->callHook('rendered');
     }
 
     protected function getPackageProviders($app): array